vulnerable-app/requirements.txt
certifi==2026.2.25
cffi==2.0.0
chardet==3.0.4
click==8.3.1
cryptography==38.0.1
Flask==2.0.1
idna==2.7
itsdangerous==2.2.0
Jinja2==3.1.6
MarkupSafe==3.0.3
pycparser==3.0
PyYAML==5.1
requests==2.20.0
urllib3==1.24.3
Werkzeug==3.1.7
